sys-apps/busybox-nix-sandbox-shell: add a llvm-libunwind hack
In Issue #28 Sora noticed that pon clang profiles like
`default/linux/amd64/17.1/systemd/clang/merged-usr` package
`busybox-nix-sandbox-shell` fails to link as:

    ld.lld: error: unable to find library -l:libunwind.a

This library is added by `clang` driver itself and is a requirement to
link any executables. For dynamic binaries it happens to work as
`libunwind.so` is already present in the system one way or another.

But for static linkage (with `-static`) `libunwind.a` is not present by
default. Ideally `sys-libs/gcc[static-libs]` should provides us
something that works. Alas.

The workaround is to pull in explicitly the `DEPEND` as:

    llvm-libunwind? ( sys-libs/llvm-libunwind[static-libs] )

It would be fine if `libc.a` and `libunwind.a` together were enough.
Unfortunately `sys-libs/gcc` as of today is built by `gcc` and that
generates a few `libgcc.a` / `libgcc-s.so` calls related to
`long double` (TF) type in `printf()` implementation: `__multf3`,
`__addtf3` and similar.

`busybox-nix-sandbox-shell` does not use any of `long double`
arithmetic.

I stubbed out all of used symbols that linker complained about with
`libtfhack.a` library that provides all these `TF`-related helpers.

The final result could fetch channel and build simple packages.

Reported-by: Sora Lee
Bug: https://bugs.gentoo.org/892956
Closes: https://github.com/trofi/nix-guix-gentoo/issues/28