dev-scheme/guile-zlib: fix `zlib` library loading
In https://github.com/trofi/nix-guix-gentoo/issues/23#issuecomment-1965152821
jaccarmac reported that `zlib` fails to load on `::gentoo`. The
reporoducer:

    $ guile --fresh-auto-compile -c '(use-modules (zlib))'
    ...
    zlib.scm:416:25: Wrong type to apply: #f

The bug happens because `guile-zlib` ebuild forced the path to point to
`/lib` (to avoid user ldscript breakage). Not that ldscript was removed
we are broken as `/lib` does not contain `libz.so` anymore.

Reported-by: jaccarmac